Resolve MUWM-4018
