+++ myuw/resources/mailman/file/__mock_key__/admin/v1.0/uwnetid/available/_uwnetid_multi_cse154a_au17
@@ -1,0 +1,1 @@
+{"Key":"multi_cse154a_au17","Available":"True"}
