+++ myuw/resources/sws/file/student/v5/enrollment/2017_autumn_9136CCB8F66711D5BE060004AC494001.json
@@ -1,0 +1,41 @@
+{
+    "ClassLevel": "NON_MATRIC",
+    "FullName": "Average,Jake Joseph",
+    "HonorsProgram": false,
+    "LeaveEndQuarter": 0,
+    "LeaveEndYear": 0,
+    "Majors": [
+                 {
+                    "Abbreviation": "UPCOM",
+                    "Campus": "Tacoma",
+                    "DegreeLevel": 2,
+                    "DegreeName": "ASD",
+                    "DegreeType": 7,
+                    "FullName": "UPCOM (Tacoma Campus)",
+                    "MajorName": "UPCOM",
+                    "Pathway": 0,
+                    "ShortName": "UPCOM"
+                }],
+    "Minors":[],
+    "PendingClassChange": false,
+    "PendingHonorsChange": false,
+    "PendingMajorChange": false,
+    "PendingResidentChange": false,
+    "PendingSpecialProgramChange": false,
+    "Person": {
+        "Href": "/student/v5/person/9136CCB8F66711D5BE060004AC494001.json",
+        "Name": "Average,Jake Joseph",
+        "RegID": "9136CCB8F66711D5BE060004AC494001"
+    },
+    "QtrGradePoints": 30,
+    "QtrGradedAttmp": 10,
+    "QtrNonGrdEarned": 2,
+    "RegID": "9136CCB8F66711D5BE060004AC494001",
+    "Registrations": [
+    ],
+    "Term": {
+        "Href": "/student/v5/term/2017,autumn.json",
+        "Quarter": "autumn",
+        "Year": "2017"
+    }
+}
