+++ myuw/resources/sws/file/student/v5/registration.json_curriculum_abbreviation_CSE_instructor_reg_id__course_number_154_verbose_true_year_2017_quarter_autumn_is_active_true_section_id_AC_transcriptable_course_all
@@ -3,7 +3,7 @@
         "ChangedSinceDate": "",
         "CourseNumber": "154",
         "CurriculumAbbreviation": "CSE",
-        "Href": "/student/v5/registration.json?year=2017&quarter=autumn&curriculum_abbreviation=CSE&course_number=154&section_id=AB&reg_id=&is_active=&instructor_reg_id=&verbose=true&changed_since_date=&transcriptable_course=all&page_start=&page_size=10",
+        "Href": "/student/v5/registration.json?year=2017&quarter=autumn&curriculum_abbreviation=CSE&course_number=154&section_id=AC&reg_id=&is_active=&instructor_reg_id=&verbose=true&changed_since_date=&transcriptable_course=all&page_start=&page_size=10",
         "InstructorRegID": "",
         "IsActive": false,
         "PageSize": "10",
@@ -24,8 +24,8 @@
             "AccessDateRangeEnd": "",
             "AccessDateRangeStart": "",
             "Auditor": false,
-            "Credits": "0",
-            "DuplicateCode": "A",
+            "Credits": "5",
+            "DuplicateCode": "",
             "EducationUnitType": null,
             "EndDate": "",
             "FeeBaseType": "",
@@ -34,25 +34,25 @@
             "GradeDocumentID": "",
             "GradingSystem": null,
             "HonorsCourse": false,
-            "Href": "/student/v5/registration/2017,autumn,CSE,154,AB,9136CCB8F66711D5BE060004AC494003,.json",
+            "Href": "/student/v5/registration/2017,autumn,CSE,154,AC,9136CCB8F66711D5BE060004AC494002,.json",
             "Instructor": null,
             "IsActive": true,
             "IsCredit": true,
             "IsIndependentStart": false,
             "Metadata": "RegistrationSourceLocation=SDB;",
             "Person": {
-                "Href": "/student/v5/person/9136CCB8F66711D5BE060004AC494003.json",
-                "Name": "Average Student",
-                "RegID": "9136CCB8F66711D5BE060004AC494003"
+                "Href": "/student/v5/person/9136CCB8F66711D5BE060004AC494002.json",
+                "Name": "Average S2dent",
+                "RegID": "9136CCB8F66711D5BE060004AC494002"
             },
             "RepeatCourse": false,
-            "RepositoryTimeStamp": "8/31/2017 12:42:18 PM",
-            "RequestDate": "2017-08-31",
+            "RepositoryTimeStamp": "8/23/2017 8:24:13 AM",
+            "RequestDate": "2017-07-26",
             "RequestStatus": "ADDED TO CLASS",
             "Section": {
                 "CourseNumber": "154",
                 "CurriculumAbbreviation": "CSE",
-                "Href": "/student/v5/course/2017,autumn,CSE,154/AB.json",
+                "Href": "/student/v5/course/2017,autumn,CSE,154/AC.json",
                 "Quarter": "autumn",
                 "SectionID": "A",
                 "Year": 2017
