+++ myuw/resources/sws/file/student/v5/registration.json_curriculum_abbreviation_CSE_instructor_reg_id__course_number_154_verbose_true_year_2017_quarter_autumn_is_active_true_section_id_A_transcriptable_course_all
@@ -24,7 +24,7 @@
             "AccessDateRangeEnd": "",
             "AccessDateRangeStart": "",
             "Auditor": false,
-            "Credits": "5",
+            "Credits": "3",
             "DuplicateCode": "A",
             "EducationUnitType": null,
             "EndDate": "",
@@ -65,7 +65,7 @@
             "AccessDateRangeEnd": "",
             "AccessDateRangeStart": "",
             "Auditor": false,
-            "Credits": "5",
+            "Credits": "4",
             "DuplicateCode": "",
             "EducationUnitType": null,
             "EndDate": "",
@@ -85,6 +85,47 @@
                 "Href": "/student/v5/person/9136CCB8F66711D5BE060004AC494002.json",
                 "Name": "Average S2dent",
                 "RegID": "9136CCB8F66711D5BE060004AC494002"
+            },
+            "RepeatCourse": false,
+            "RepositoryTimeStamp": "8/23/2017 8:24:13 AM",
+            "RequestDate": "2017-07-26",
+            "RequestStatus": "ADDED TO CLASS",
+            "Section": {
+                "CourseNumber": "154",
+                "CurriculumAbbreviation": "CSE",
+                "Href": "/student/v5/course/2017,autumn,CSE,154/A.json",
+                "Quarter": "autumn",
+                "SectionID": "A",
+                "Year": 2017
+            },
+            "StartDate": "",
+            "VariableCredit": false,
+            "WritingCourse": false
+        },
+        {
+            "AccessDateRangeEnd": "",
+            "AccessDateRangeStart": "",
+            "Auditor": false,
+            "Credits": "5",
+            "DuplicateCode": "",
+            "EducationUnitType": null,
+            "EndDate": "",
+            "FeeBaseType": "",
+            "Grade": "X",
+            "GradeDate": "",
+            "GradeDocumentID": "",
+            "GradingSystem": null,
+            "HonorsCourse": false,
+            "Href": "/student/v5/registration/2017,autumn,CSE,154,A,9136CCB8F66711D5BE060004AC494001,.json",
+            "Instructor": null,
+            "IsActive": true,
+            "IsCredit": true,
+            "IsIndependentStart": false,
+            "Metadata": "RegistrationSourceLocation=SDB;",
+            "Person": {
+                "Href": "/student/v5/person/9136CCB8F66711D5BE060004AC494001.json",
+                "Name": "Average S2dent",
+                "RegID": "9136CCB8F66711D5BE060004AC494001"
             },
             "RepeatCourse": false,
             "RepositoryTimeStamp": "8/23/2017 8:24:13 AM",
