+++ myuw/test/api/test_instructor_schedule.py
@@ -1,10 +1,8 @@
+import json
 from django.test.utils import override_settings
 from myuw.test.api import require_url, MyuwApiTest
-from myuw.test.views.lti import get_lti_request, MyuwLTITest
 from restclients_core.exceptions import DataFailureException
-from myuw.views.api.instructor_schedule import (
-    InstScheCurQuar, InstSect, InstSectionDetails, LTIInstSectionDetails)
-import json
+from myuw.views.api.instructor_schedule import InstScheCurQuar, InstSect
 from myuw.dao.instructor_schedule import (
     get_current_quarter_instructor_schedule)
 from myuw.test import get_request, get_request_with_user, get_request_with_date
@@ -205,82 +203,3 @@ class TestInstructorSection(MyuwApiTest):
                          '2017_autumn_EDC_I_552_A')
         self.assertEqual(data['sections'][0]['curriculum_abbr'],
                          'EDC&I')
-
-
-class TestInstructorSectionDetails(MyuwApiTest):
-    def test_bill_section(self):
-        request = get_request()
-        get_request_with_user('bill', request)
-
-        section_id = '2013,spring,ESS,102/A'
-        resp = InstSectionDetails().get(request, section_id=section_id)
-        self.assertEqual(resp.status_code, 200)
-
-        data = json.loads(resp.content)
-        self.assertEqual(len(data['sections']), 1)
-        self.assertEqual(
-            data['sections'][0]['limit_estimate_enrollment'], 15)
-        self.assertEqual(
-            data['sections'][0]['final_exam']['latitude'],
-            47.656645546715)
-        self.assertEqual(data['sections'][0]['canvas_url'],
-                         'https://canvas.uw.edu/courses/149651')
-        self.assertEqual(
-            len(data['sections'][0]['grade_submission_delegates']), 1)
-
-        self.assertGreater(len(data['related_terms']), 3)
-        self.assertEqual(data['related_terms'][
-            len(data['related_terms']) - 3]['quarter'], 'Spring')
-        self.assertEqual(data['related_terms'][5]['year'], 2013)
-
-    def test_invalid_section(self):
-        request = get_request()
-        get_request_with_user('bill', request)
-
-        section_id = ''
-        resp = InstSectionDetails().get(request, section_id=section_id)
-        self.assertEqual(resp.status_code, 400)
-
-        section_id = '12345'
-        resp = InstSectionDetails().get(request, section_id=section_id)
-        self.assertEqual(resp.status_code, 400)
-
-
-@override_settings(BLTI_AES_KEY=b"11111111111111111111111111111111",
-                   BLTI_AES_IV=b"1111111111111111")
-class TestLTIInstructorSectionDetails(MyuwLTITest):
-    def test_bill_section(self):
-        request = get_lti_request()
-
-        section_id = '2013-spring-ESS-102-A'
-        resp = LTIInstSectionDetails().get(request, section_id=section_id)
-
-        self.assertEqual(resp.status_code, 200)
-
-        data = json.loads(resp.content)
-        self.assertEqual(len(data['sections']), 1)
-        self.assertEqual(
-            data['sections'][0]['limit_estimate_enrollment'], 15)
-        self.assertEqual(
-            data['sections'][0]['final_exam']['latitude'],
-            47.656645546715)
-        self.assertEqual(data['sections'][0]['canvas_url'],
-                         'https://canvas.uw.edu/courses/149651')
-        self.assertEqual(
-            len(data['sections'][0]['grade_submission_delegates']), 1)
-
-        self.assertGreater(len(data['related_terms']), 3)
-        self.assertEqual(data['related_terms'][
-            len(data['related_terms']) - 3]['quarter'], 'Spring')
-        self.assertEqual(data['related_terms'][5]['year'], 2013)
-
-    def test_invalid_section(self):
-        request = get_lti_request()
-
-        section_id = ''
-        resp = LTIInstSectionDetails().get(request, section_id=section_id)
-        self.assertEqual(resp.status_code, 403)
-
-        section_id = '12345'
-        resp = LTIInstSectionDetails().get(request, section_id=section_id)
-        self.assertEqual(resp.status_code, 403)
