+++ myuw/views/api/instructor_schedule.py
@@ -25,9 +25,7 @@ from myuw.dao.instructor_schedule import (
     get_instructor_section, get_primary_section, check_section_instructor)
 from myuw.dao.library import get_subject_guide_by_section
 from myuw.dao.mailman import get_section_email_lists
-from myuw.dao.pws import (
-    get_person_of_current_user, get_person_by_regid as get_pws_person_by_regid,
-    get_url_key_for_regid)
+from myuw.dao.pws import get_person_of_current_user
 from myuw.dao.registration import get_active_registrations_for_section
 from myuw.dao.term import (
     get_current_quarter, is_past, is_future, get_previous_number_quarters,
@@ -460,220 +458,3 @@ class InstSect(ProtectedAPI):
             return self.make_http_resp(timer, request, section_id)
         except Exception:
             return handle_exception(logger, timer, traceback)
-
-
-class OpenInstSectionDetails(OpenAPI):
-    """
-    Performs actions on resource at
-    /api/v1/instructor_section/<year>,<quarter>,<curriculum>,
-        <course_number>,<course_section>?
-    """
-    def validate_section_id(self, request, section_id):
-        return section_id
-
-    def is_authorized_for_section(self, request, schedule):
-        raise NotSectionInstructorException()
-
-    def make_http_resp(self, timer, request, section_id):
-        """
-        @return instructor schedule data in json format
-            status 404: no schedule found (teaching no courses)
-        """
-        self.processed_primary = False
-        self.person = get_person_of_current_user()
-
-        try:
-            section_id = self.validate_section_id(request, section_id)
-        except NotSectionInstructorException:
-            return not_instructor_error()
-
-        schedule = get_instructor_section(self.person, section_id,
-                                          include_registrations=True,
-                                          include_linked_sections=True)
-
-        try:
-            self.is_authorized_for_section(request, schedule)
-        except NotSectionInstructorException:
-            return not_instructor_error()
-
-        self.term = schedule.term
-        resp_data = load_schedule(request, schedule,
-                                  section_callback=self.per_section_data)
-
-        self.add_linked_section_data(resp_data)
-        log_success_response(logger, timer)
-
-        return self.json_response(resp_data)
-
-    def add_linked_section_data(self, resp_data):
-        section_types = {}
-        sections_for_user = {}
-        all_types = set()
-        for section in resp_data["sections"][1:]:
-            section_id = section["section_id"]
-            section_type = section["section_type"]
-            section_types[section_id] = section_type
-            all_types.add(section_type)
-            for registration in section["registrations"]:
-                if registration not in sections_for_user:
-                    sections_for_user[registration] = []
-                sections_for_user[registration].append(section_id)
-
-        for registration in resp_data["sections"][0]["registrations"]:
-            regid = registration["regid"]
-            registration["linked_sections"] = []
-            if regid in sections_for_user:
-                user_types = {}
-                for section in sections_for_user[regid]:
-                    section_type = section_types[section]
-
-                    if section_type not in user_types:
-                        user_types[section_type] = []
-
-                    user_types[section_type].append(section)
-
-                for section_type in sorted(all_types):
-                    # These are for sorting
-                    if section_type in user_types:
-                        sort_string = ",".join(user_types[section_type])
-                        registration["linked_sections"].append(
-                            {'type': section_type,
-                             'sections': user_types[section_type]})
-                    else:
-                        sort_string = "ZZ"
-                        registration["linked_sections"].append(
-                            {'type': section_type,
-                             'sections': [""]})
-                    registration["linked_type_"+section_type] = sort_string
-
-        types_list = list(set(section_types.values()))
-        resp_data["sections"][0]["linked_types"] = types_list
-
-    def per_section_data(self, section, section_data):
-        # We don't want to fetch all this data a second time in for
-        # secondary sections
-        if self.processed_primary:
-            registrations = []
-            for registration in section.registrations:
-                registrations.append(registration.person.uwregid)
-
-            section_data["registrations"] = registrations
-            return
-
-        self.processed_primary = True
-        registrations = {}
-        name_threads = {}
-        enrollment_threads = {}
-
-        for registration in section.registrations:
-            person = registration.person
-            regid = person.uwregid
-
-            name_email_thread = ThreadWithResponse(target=self.get_person_info,
-                                                   args=(person,))
-            enrollment_thread = ThreadWithResponse(target=self.get_enrollments,
-                                                   args=(person,))
-
-            name_threads[regid] = name_email_thread
-            enrollment_threads[regid] = enrollment_thread
-            name_email_thread.start()
-            enrollment_thread.start()
-
-            registrations[regid] = {
-                'full_name': person.display_name,
-                'netid': person.uwnetid,
-                'regid': person.uwregid,
-                'student_number': person.student_number,
-                'credits': registration.credits,
-                'is_auditor': registration.is_auditor,
-                'class': person.student_class,
-                'email': person.email1,
-                'url_key': get_url_key_for_regid(person.uwregid),
-            }
-
-        registration_list = []
-        for regid in name_threads:
-            thread = name_threads[regid]
-            thread.join()
-            registrations[regid]["name"] = thread.response["name"]
-            registrations[regid]["surname"] = thread.response["surname"]
-            registrations[regid]["first_name"] = thread.response["first_name"]
-            registrations[regid]["email"] = thread.response["email"]
-
-            thread = enrollment_threads[regid]
-            thread.join()
-            if thread.response:
-                registrations[regid]["majors"] = thread.response["majors"]
-                registrations[regid]["class"] = thread.response["class"]
-
-                code = get_code_for_class_level(thread.response["class"])
-                registrations[regid]['class_code'] = code
-
-            registration_list.append(registrations[regid])
-        section_data["registrations"] = registration_list
-
-    def get_person_info(self, person):
-        sws_person = get_person_by_regid(person.uwregid)
-        try:
-            fname = sws_person.first_name.title()
-        except AttributeError:
-            fname = ""
-        return {"name": fname,
-                "surname": sws_person.last_name.title(),
-                "first_name": sws_person.first_name.title(),
-                "email": sws_person.email}
-
-    def get_enrollments(self, person):
-        regid = person.uwregid
-
-        enrollments = get_enrollment_by_regid_and_term(regid, self.term)
-        majors = []
-        for major in enrollments.majors:
-            majors.append(major.json_data())
-
-        return {"majors": majors, "class": enrollments.class_level}
-
-    def get(self, request, *args, **kwargs):
-        """
-        GET returns 200 with a specific term instructor schedule
-        @return course schedule data in json format
-                status 400: invalid section_id
-                status 403: not authorized to view section details
-                status 404: no schedule found (not registered)
-                status 543: data error
-        """
-        section_id = kwargs.get('section_id')
-        timer = Timer()
-        try:
-            return self.make_http_resp(timer, request, section_id)
-        except Exception as ex:
-            return handle_exception(logger, timer, traceback)
-
-
-@method_decorator(login_required, name='dispatch')
-class InstSectionDetails(OpenInstSectionDetails):
-    def is_authorized_for_section(self, request, schedule):
-        try:
-            check_section_instructor(schedule.sections[0], schedule.person)
-        except IndexError:
-            pass
-
-
-@method_decorator(blti_admin_required, name='dispatch')
-class LTIInstSectionDetails(OpenInstSectionDetails):
-    def is_authorized_for_section(self, request, schedule):
-        pass
-
-    def validate_section_id(self, request, section_id):
-        blti_data = BLTI().get_session(request)
-        authorized_sections = blti_data.get('authorized_sections', [])
-
-        if section_id not in authorized_sections:
-            raise NotSectionInstructorException
-
-        (sws_section_id, instructor_regid) = sws_section_label(section_id)
-
-        if instructor_regid is not None:
-            self.person = get_pws_person_by_regid(instructor_regid)
-
-        return sws_section_id
