+++ setup.py
@@ -36,7 +36,7 @@ setup(
         'AuthZ-Group',
         'python-binary-memcached',
         'UW-RestClients-Core==0.9.1',
-        'UW-RestClients-SWS<2.0,>=1.4.4',
+        'UW-RestClients-SWS>=1.4.6.0a1,<2',
         'UW-RestClients-PWS<1.0,>=0.6',
         'UW-RestClients-HFS<1.0',
         'UW-RestClients-GWS<1.0',
